backend/pyproject.toml
[project]
name = "booza-shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.0",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "ruff>=0.15.0",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "sqlalchemy>=2.0.46",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
env_files = [".env.test"]
python_classes = ["Test*"]
python_functions = ["test_*"]
